add taginfo file to let OSMers know which tags we use
Taginfo is a very useful project to learn about tags in OSM, and you can use it to know what the tags are used to.

this PR add a taginfo file for our project, so it can be displayed in the pages for the tags we use.

The documentation for the file format can be found here : https://wiki.openstreetmap.org/wiki/Taginfo/Projects