add new output format
add the possibility to dump json stream file

naming the ouput file "*.jsonl" (or ".jsonl.gz") will dump the cosmogony
file as jsons stream, each line being a zone as json.

it makes it way ligher to read the file (it's not mandatory to load the
whole file in memory).

For easier cosmogony use, several reading methods have been provided:

* `load_cosmogony_from_file` -> deduce the file format from extension
and load the whole cosmogony
* `read_zones_from_file` -> decude the file format and return an
iterator on the zones

* `read_zones`: stream zones from a buffered reader. Can be a bit more
efficient if the cosmogony format is known.