include the libpostal's rules in the cosmogony binary
This way the cosmogony is self sufficient, no need for additional rules.
It will makes it easier to package it.

It is a breaking change as now there is no more the `--libpostal` cli
parameter

:warning:
The `include_dir` macro cannot cannot watch for changes to rebuild,
so if you only change the libpostal rules, you need either to clean the
repo before building
or just touch thie `src/zone_typer.rs` file to trigger a reimport.